refactor(tarana-eats): restructure data into modular files for better maintainability
Split taranaEatsData.ts into modular files under /data directory:
- /types.ts for type definitions
- /helpers.ts for utility functions
- /restaurants.ts for restaurant data
- /menus/ directory for individual restaurant menus
- /aiData.ts for AI recommendation data
- /index.ts for central exports

Maintain backward compatibility while improving organization and scalability
