+++ src/app/tarana-eats/data/menus/goodShepherdCafe.ts
@@ -1,0 +1,68 @@
+// ====================================================================
+// GOOD SHEPHERD CAFE MENU DATA
+// ====================================================================
+
+import { FullMenu } from "../types";
+
+export const goodShepherdCafeMenu: FullMenu = {
+  Breakfast: [],
+  Lunch: [],
+  Dinner: [],
+  Snacks: [
+    {
+      name: "Long Black (Coffee)",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 125,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "Cafe Latte (Coffee)",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 145,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "Cappuccino (Coffee)",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 145,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "Caramel Macchiato (Coffee)",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 185,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "White Mocha Latte (Coffee)",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 185,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "Shepherd's Latte",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 210,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "Orange Espresso",
+      description: "Our coffee is crafted from locally sourced beans handpicked from the farms in Atok, Benguet - bringing the taste of the Cordilleras to every sip.",
+      price: 210,
+      image: "/images/placeholders/coffee.png",
+    },
+    {
+      name: "Pineapple Hibiscus",
+      description: "No description available — yet.",
+      price: 160,
+      image: "/images/placeholders/juice.png",
+    },
+    {
+      name: "Calamansi Juice",
+      description: "No description available — yet.",
+      price: 160,
+      image: "/images/placeholders/juice.png",
+    },
+  ],
+  Drinks: [],
+};
