+++ src/app/tarana-eats/data/menus/kFlavorsBuffet.ts
@@ -1,0 +1,143 @@
+// ====================================================================
+// K-FLAVORS BUFFET MENU DATA
+// ====================================================================
+
+import { FullMenu } from "../types";
+
+export const kFlavorsBuffetMenu: FullMenu = {
+  Breakfast: [
+    {
+      name: "Buffet (M-TH, No beef/cheese/ice cream)",
+      description: "NO beef, cheese, ice cream",
+      price: 399,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (M-TH, With beef/cheese/ice cream)",
+      description: "with beef, cheese, 1 scoop of ice cream",
+      price: 450,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (M-TH, With beef/cheese/ice cream/unli iced tea)",
+      description: "with beef, cheese, 1 scoop of ice cream, unli iced tea",
+      price: 470,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, No beef/cheese/ice cream)",
+      description: "NO beef, cheese, ice cream",
+      price: 450,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, With beef/cheese/ice cream)",
+      description: "with beef, cheese, 1 scoop of ice cream",
+      price: 470,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, With beef/cheese/ice cream/unli iced tea)",
+      description: "with beef, cheese, 1 scoop of ice cream, unli iced tea",
+      price: 499,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    }
+  ],
+  Lunch: [
+    {
+      name: "Buffet (M-TH, No beef/cheese/ice cream)",
+      description: "NO beef, cheese, ice cream",
+      price: 399,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (M-TH, With beef/cheese/ice cream)",
+      description: "with beef, cheese, 1 scoop of ice cream",
+      price: 450,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (M-TH, With beef/cheese/ice cream/unli iced tea)",
+      description: "with beef, cheese, 1 scoop of ice cream, unli iced tea",
+      price: 470,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, No beef/cheese/ice cream)",
+      description: "NO beef, cheese, ice cream",
+      price: 450,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, With beef/cheese/ice cream)",
+      description: "with beef, cheese, 1 scoop of ice cream",
+      price: 470,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, With beef/cheese/ice cream/unli iced tea)",
+      description: "with beef, cheese, 1 scoop of ice cream, unli iced tea",
+      price: 499,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    }
+  ],
+  Dinner: [
+    {
+      name: "Buffet (M-TH, No beef/cheese/ice cream)",
+      description: "NO beef, cheese, ice cream",
+      price: 399,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (M-TH, With beef/cheese/ice cream)",
+      description: "with beef, cheese, 1 scoop of ice cream",
+      price: 450,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (M-TH, With beef/cheese/ice cream/unli iced tea)",
+      description: "with beef, cheese, 1 scoop of ice cream, unli iced tea",
+      price: 470,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, No beef/cheese/ice cream)",
+      description: "NO beef, cheese, ice cream",
+      price: 450,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, With beef/cheese/ice cream)",
+      description: "with beef, cheese, 1 scoop of ice cream",
+      price: 470,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    },
+    {
+      name: "Buffet (F-S, With beef/cheese/ice cream/unli iced tea)",
+      description: "with beef, cheese, 1 scoop of ice cream, unli iced tea",
+      price: 499,
+      image: "https://www.facebook.com/photo.php?fbid=660691786987731&set=pb.100091407458861.-2207520000&type=3",
+    }
+  ],
+  Snacks: [
+    {
+      name: "Unlimited Iced Tea",
+      description: "Bottomless iced tea with buffet",
+      price: 20,
+      image: "/images/placeholders/drink.png",
+    },
+    {
+      name: "Soft Drinks",
+      description: "Coca-cola products",
+      price: 25,
+      image: "/images/placeholders/drink.png",
+    },
+    {
+      name: "Bottled Water",
+      description: "Mineral water",
+      price: 15,
+      image: "/images/placeholders/drink.png",
+    }
+  ],
+  Drinks: [],
+};
