+++ src/app/tarana-eats/data/menus/myeongDongJjigae.ts
@@ -1,0 +1,38 @@
+// ====================================================================
+// MYEONG DONG JJIGAE RESTAURANT MENU DATA
+// ====================================================================
+
+import { FullMenu } from "../types";
+
+export const myeongDongJjigaeMenu: FullMenu = {
+  Breakfast: [],
+  Lunch: [
+    {
+      name: "Myeongdong Set A",
+      description: "Buffet Only",
+      price: 399,
+      image: "https://www.facebook.com/photo.php?fbid=122177289860112650&set=pb.61553379521587.-2207520000&type=3",
+    },
+    {
+      name: "Myeongdong Set B",
+      description: "Buffet with Samgyeopsal",
+      price: 599,
+      image: "https://www.facebook.com/photo?fbid=122177289872112650&set=pb.61553379521587.-2207520000",
+    },
+    {
+      name: "Myeongdong Set C",
+      description: "Buffet with Hotpot/Jjigae",
+      price: 599,
+      image: "https://www.facebook.com/photo?fbid=122177289854112650&set=pb.61553379521587.-2207520000",
+    },
+    {
+      name: "Myeongdong Set D",
+      description: "All-in-Buffet with Samgyeopsal and Jjigae",
+      price: 699,
+      image: "https://www.facebook.com/photo?fbid=122177289866112650&set=pb.61553379521587.-2207520000",
+    },
+  ],
+  Dinner: [],
+  Snacks: [],
+  Drinks: [],
+};
