+++ src/app/tarana-eats/data/menus/ohMyGulay.ts
@@ -1,0 +1,111 @@
+// ====================================================================
+// OH MY GULAY MENU DATA
+// ====================================================================
+
+import { FullMenu } from "../types";
+
+export const ohMyGulayMenu: FullMenu = {
+  Breakfast: [],
+  Lunch: [
+    {
+      name: "Bonifacio's Waldofesto",
+      description: "Sliced apples, celery, potatoes, egg slices, nuts & garlic-mayo pesto",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513120.jpg??width=800",
+    },
+    {
+      name: "KKK Sliders",
+      description: "Two veggie burgers & salad",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513121.jpg??width=800",
+    },
+    {
+      name: "Pancit Gulay Ni Pepe",
+      description: "Egg noodles, tofu & vegetables in season",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513127.jpg??width=800",
+    },
+    {
+      name: "Walastik Bistek",
+      description: "Vegemeat, tofu, potatoes, onions & quail eggs",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513129.jpg??width=800",
+    },
+    {
+      name: "Tandang Sora Parmigiana",
+      description: "Breaded eggplant over whole wheat bread, special red sauce, pesto & cheese",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513122.jpg??width=800",
+    },
+    {
+      name: "CESAR ASAR",
+      description: "Vegetarian Ceasar Salad with toasted garlic bread",
+      price: 200,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/32477040.jpg??width=800",
+    },
+    {
+      name: "Bandilang Pula",
+      description: "Mountain rice, vegetables in season, mongolian sauce, tofu, omelette",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513128.jpg??width=800",
+    },
+    {
+      name: "Anak Ng Putanesca",
+      description: "No description available — yet.",
+      price: 220,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513123.jpg??width=800",
+    },
+  ],
+  Dinner: [
+    {
+      name: "Bonifacio's Waldofesto",
+      description: "Sliced apples, celery, potatoes, egg slices, nuts & garlic-mayo pesto",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513120.jpg??width=800",
+    },
+    {
+      name: "KKK Sliders",
+      description: "Two veggie burgers & salad",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513121.jpg??width=800",
+    },
+    {
+      name: "Pancit Gulay Ni Pepe",
+      description: "Egg noodles, tofu & vegetables in season",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513127.jpg??width=800",
+    },
+    {
+      name: "Walastik Bistek",
+      description: "Vegemeat, tofu, potatoes, onions & quail eggs",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513129.jpg??width=800",
+    },
+    {
+      name: "Tandang Sora Parmigiana",
+      description: "Breaded eggplant over whole wheat bread, special red sauce, pesto & cheese",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513122.jpg??width=800",
+    },
+    {
+      name: "CESAR ASAR",
+      description: "Vegetarian Ceasar Salad with toasted garlic bread",
+      price: 200,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/32477040.jpg??width=800",
+    },
+    {
+      name: "Bandilang Pula",
+      description: "Mountain rice, vegetables in season, mongolian sauce, tofu, omelette",
+      price: 225,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513128.jpg??width=800",
+    },
+    {
+      name: "Anak Ng Putanesca",
+      description: "No description available — yet.",
+      price: 220,
+      image: "https://images.deliveryhero.io/image/fd-ph/products/5513123.jpg??width=800",
+    },
+  ],
+  Snacks: [],
+  Drinks: [],
+};
