+++ src/app/tarana-eats/data/menus/ujiMatchaCafe.ts
@@ -1,0 +1,86 @@
+// ====================================================================
+// UJI-MATCHA CAFE MENU DATA
+// ====================================================================
+
+import { FullMenu } from "../types";
+
+export const ujiMatchaCafeMenu: FullMenu = {
+  Breakfast: [],
+  Lunch: [],
+  Dinner: [],
+  Snacks: [
+    {
+      name: "Pure Uji-Matcha",
+      description: "Matcha + Water + Sweetener",
+      price: 110,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974537.jpg??width=800",
+    },
+    {
+      name: "Matcha Latte",
+      description: "Matcha + Milk + Sweetener",
+      price: 155,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974542.jpg??width=800",
+    },
+    {
+      name: "Spanish Matcha Latte",
+      description: "Matcha + Milk + Condense Milk",
+      price: 155,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974543.jpg??width=800",
+    },
+    {
+      name: "Honey Matcha Latte",
+      description: "Matcha + Milk + Wildflower Honey",
+      price: 160,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/53768974.jpg??width=800",
+    },
+    {
+      name: "Dirty Matcha Latte",
+      description: "Matcha + Milk + Biscoff",
+      price: 165,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974552.jpg??width=800",
+    },
+    {
+      name: "Strawberry Matcha Latte (Iced Only)",
+      description: "Matcha + Milk + Strawberry",
+      price: 155,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974553.jpg??width=800",
+    },
+    {
+      name: "Blueberry Matcha Latte (Iced Only)",
+      description: "Matcha + Milk + Blueberry",
+      price: 155,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974554.jpg??width=800",
+    },
+    {
+      name: "Sweet Black (Espresso Based)",
+      description: "2 shots Espresso + Water + Sweetener",
+      price: 100,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974574.jpg??width=800",
+    },
+    {
+      name: "Spanish Latte (Espresso Based)",
+      description: "1 or 2 shots Espresso + Milk + Condense Milk",
+      price: 140,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974577.jpg??width=800",
+    },
+    {
+      name: "Mocha Latte (Espresso Based)",
+      description: "1 shot Espresso + Milk + Salted Caramel",
+      price: 140,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974579.jpg??width=800",
+    },
+    {
+      name: "Strawberry Milk (Non-Matcha)",
+      description: "Fresh milk infused with the natural sweetness of strawberries, delivers a fruity and refreshing flavor",
+      price: 125,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/39974581.jpg??width=800",
+    },
+    {
+      name: "Houjicha Latte",
+      description: "Houjicha + Milk + Sweetener",
+      price: 160,
+      image: "https://images.deliveryhero.io/image/fd-ph/Products/64742086.jpg??width=800",
+    }
+  ],
+  Drinks: [],
+};
